Update and rename bdtext to bdt.sql

+++ bdt.sql
@@ -2,8 +2,8 @@ CREATE TABLE `test`.`etepac`(
  `id` INT NOT NULL AUTO_INCREMENT,
  `aluno` VARCHAR(50) NOT NULL,
  `disciplina` VARCHAR(20) NOT NULL,
- `nota1` FLOAT NOT NULL,
- `nota2` FLOAT NOT NULL,
- `nota3` FLOAT NOT NULL,
+ `nota1` int NOT NULL,
+ `nota2` int NOT NULL,
+ `nota3` int NOT NULL,
  PRIMARY KEY(`id`)
 ) ENGINE = INNODB;
